+++ src/index.ts
@@ -2,4 +2,3 @@
  * Copyright (c) 2026 Interop Alliance. All rights reserved.
  */
 export { CapabilityAgent } from './CapabilityAgent.js'
-export type { VerificationKeyDescriptor } from './CapabilityAgent.js'